rounded float to int to prevent error

+++ facerecognition/face.py
@@ -29,8 +29,9 @@ def crop(image, x, y, w, h):
     """
     crop_height = int((config.FACE_HEIGHT / float(config.FACE_WIDTH)) * w)
     midy = y + h / 2
-    y1 = max(0, midy - crop_height / 2)
-    y2 = min(image.shape[0] - 1, midy + crop_height / 2)
+    y1 = int(max(0, midy - crop_height / 2))
+    y2 = int(min(image.shape[0] - 1, midy + crop_height / 2))
+
     return image[y1:y2, x:x + w]
